+++ specs/fixtures/basic_app_config/pages/index.vue
@@ -1,22 +1,7 @@
 <script setup lang="ts">
-import { navigateTo } from '#imports'
-
 const { locale } = useI18n()
 const localePath = useLocalePath()
 const switchLocalePath = useSwitchLocalePath()
-const localeRoute = useLocaleRoute()
-
-const category = ref({
-  title: 'Kirby',
-  slug: 'nintendo'
-})
-
-function onClick() {
-  const route = localeRoute({ name: 'user-profile', query: { foo: '1' } })
-  if (route) {
-    return navigateTo(route.fullPath)
-  }
-}
 </script>
 
 <template>
@@ -54,10 +39,6 @@ function onClick() {
           <NuxtLink :to="switchLocalePath('fr')">Français</NuxtLink>
         </li>
       </ul>
-    </section>
-    <section id="locale-route-usages">
-      <h3>localeRoute</h3>
-      <button @click="onClick">Show profile</button>
     </section>
   </div>
 </template>
